make a new auth module in types that contains UserInfo and friends
