+++ types/Cargo.toml
@@ -12,6 +12,7 @@ path = "src/lib.rs"
 serde = { version = "1.0", features = ["derive"] }
 serde_with = "1.9"
 serde_plain = "0.3"
+log = "0.4"
 serde_json = "1"
 tokio-stream = "0.1"
 anyhow = "1.0"
