+++ types/src/user.rs
@@ -1,5 +1,4 @@
 use serde::{Deserialize, Serialize};
-use serde_with::{rust::StringWithSeparator, CommaSeparator};
 use uuid::Uuid;
 
 /// A user id tied back to AWS Cognito `sub` claim.
@@ -20,124 +19,4 @@ pub struct User {
     pub id: UserId,
     /// User-facing name for this contributor/curator
     pub display_name: String,
-}
-
-/// Auth metadata on the user making the current request.
-#[derive(PartialEq, Deserialize, Debug, async_graphql::SimpleObject)]
-pub struct UserInfo {
-    /// Unique ID for the User. Should be an AWS Cognito Sub.
-    #[serde(default, rename = "sub")]
-    pub id: Uuid,
-    email: String,
-    #[serde(default, rename = "cognito:groups")]
-    pub groups: Vec<UserGroup>,
-}
-
-// AWS ApiGateway will always encode the groups as a comma-separated
-// string, even if the client sends a JWT with an array.
-#[derive(PartialEq, Deserialize, Debug)]
-#[serde(remote = "UserInfo")]
-pub struct ApiGatewayUserInfoDef {
-    #[serde(default, rename = "sub")]
-    id: Uuid,
-    email: String,
-    #[serde(
-        default,
-        rename = "cognito:groups",
-        with = "StringWithSeparator::<CommaSeparator>"
-    )]
-    groups: Vec<UserGroup>,
-}
-
-#[derive(Deserialize)]
-pub struct ApiGatewayUserInfo(#[serde(with = "ApiGatewayUserInfoDef")] UserInfo);
-
-impl Into<UserInfo> for ApiGatewayUserInfo {
-    fn into(self) -> UserInfo {
-        self.0
-    }
-}
-
-#[derive(Eq, PartialEq, Copy, Clone, Serialize, Deserialize, Debug, async_graphql::Enum)]
-pub enum UserGroup {
-    Contributors,
-    Editors,
-}
-// // Impl FromStr and Display automatically for UserGroup, using serde.
-// // This allows us to (de)serialize lists of groups via a comma-separated string
-// // like this: "Editor,Contributor,Translator"
-serde_plain::forward_from_str_to_serde!(UserGroup);
-serde_plain::forward_display_to_serde!(UserGroup);
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn decoding_single_group_works() {
-        let res: Result<ApiGatewayUserInfo, _> = serde_json::from_str(
-            r#"
-            {
-                "cognito:groups": "Editors",
-                "cognito:username": "7c455493-8e7e-47b9-abed-5f1492eb7a9b",
-                "email": "charliemcvicker@protonmail.com",
-                "sub": "7c455493-8e7e-47b9-abed-5f1492eb7a9b"
-              }
-            "#,
-        );
-
-        assert_eq!(
-            res.unwrap().into(),
-            UserInfo {
-                id: Uuid::parse_str("7c455493-8e7e-47b9-abed-5f1492eb7a9b").unwrap(),
-                email: String::from("charliemcvicker@protonmail.com"),
-                groups: vec![UserGroup::Editors]
-            }
-        )
-    }
-
-    #[test]
-    fn decoding_many_groups_works() {
-        let res: Result<ApiGatewayUserInfo, _> = serde_json::from_str(
-            r#"
-            {
-                "cognito:groups": "Editors,Contributors",
-                "cognito:username": "7c455493-8e7e-47b9-abed-5f1492eb7a9b",
-                "email": "charliemcvicker@protonmail.com",
-                "sub": "7c455493-8e7e-47b9-abed-5f1492eb7a9b"
-              }
-            "#,
-        );
-
-        assert_eq!(
-            res.unwrap().into(),
-            UserInfo {
-                id: Uuid::parse_str("7c455493-8e7e-47b9-abed-5f1492eb7a9b").unwrap(),
-                email: String::from("charliemcvicker@protonmail.com"),
-                groups: vec![UserGroup::Editors, UserGroup::Contributors]
-            }
-        )
-    }
-
-    #[test]
-    fn decoding_no_groups_works() {
-        let res: Result<ApiGatewayUserInfo, _> = serde_json::from_str(
-            r#"
-            {
-                "cognito:username": "7c455493-8e7e-47b9-abed-5f1492eb7a9b",
-                "email": "charliemcvicker@protonmail.com",
-                "sub": "7c455493-8e7e-47b9-abed-5f1492eb7a9b"
-              }
-            "#,
-        );
-
-        assert_eq!(
-            res.unwrap().into(),
-            UserInfo {
-                id: Uuid::parse_str("7c455493-8e7e-47b9-abed-5f1492eb7a9b").unwrap(),
-                email: String::from("charliemcvicker@protonmail.com"),
-                groups: vec![]
-            }
-        )
-    }
 }
